redux file reorganizing
