+++ src/pages/admin/index.js
@@ -51,7 +51,7 @@ const AdminRoutes = (props) => {
 
 const mapStateToProps = state => {
   return {
-    events: state.pageState.events,
+    events: state.eventState.events,
     user: state.userState.user
   }
 }
