+++ src/pages/public/Login.js
@@ -13,7 +13,7 @@ import {
 
 import LoginImage from 'assets/login.svg'
 import { COLORS } from 'constants/index'
-import { setUser } from 'store/user/UserActions'
+import { setUser } from 'store/user/userActions'
 import { checkFeatureFlag } from 'utils'
 
 const styles = {
